ENH: improved smoothing algorithm
The previous version would often not succeed if
quite a few of the last entries of a to-be-smoothened
array would be 0 (e.g. integrals at the largest
distances for which the integration grid is empty).
Now the smoothing is performed after removing the
trailing zeros, and putting them back afterwards.