MAINT: avoiding unimportant numpy warnings
Previously, there were quite a few 'division by zero'
and 'invalid value encountered' warnings, due to e.g.
the (GPAW) grid starting at 0, or very low values for
the electron density sent to XC_PW92, or in the calcu-
lation of the angles when making the integration grid
in hotcent.slako. Since these warnings are unimportant
/without consequence, it is better to simple avoid them.