ENH: do not calculate both on3c triplets at once
Requiring separate calls for e.g. A-B-C and A-C-B triplets
makes this more in line with how the off3c and rep3c generators
are working. It also allows to e.g. choose different Rgrids
for the two triplets, which is beneficial if B and C have
significantly different covalent radii.