ENH: more robust by skipping high-V endpoints
With SoftConfinement the potential diverges at the
cutoff radius. So far all grid points within this
radius were taken into account, but the confinement
potential may still be very large near the endpoints,
occasionally causing the "c0[-2] < 0 and c0[-1] < 0"
assertion to fail. This is remedied by further
restricting the Veff range when needed.