ENH/BUG: more accurate poisson solver for on1cU/off2cU
The prior version did not e.g. take into account that
the rightmost boundary condition should not be 0 (l>0)
or 4*pi (l=0), due to the finite extent of the radial
integration grid. It's better to just use a grid as in
Becke and Dickson (1988), where r=0 and r=+inf are the
ghost points.

Also some superfluous OrbitalHartreePotential builds
and tail_smoothening calls are removed here.