ENH: add and check SHA256 sum for pseudo file
Since hotcent-basis and hotcent-tables may get
executed on different machines we can't check
the PP file path itself, but we can verify the
checksums to make sure that the same PP is used.
If (really) necessary, one can disable the check
by removing/commenting the entry in the YAML file.